(GH-4374) Fix Frosting argument parsing * Add argument to Forsting integration tests * Unify Scripting / Frosting Spectre.Console IRemainingArguments to CakeArguments parsing * fixes #4374
