+++ src/Cake.Frosting/Internal/Commands/DefaultCommand.cs
@@ -6,6 +6,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Cake.Cli;
+using Cake.Cli.Infrastructure;
 using Cake.Core;
 using Cake.Core.Diagnostics;
 using Cake.Core.IO;
@@ -81,31 +82,7 @@ namespace Cake.Frosting.Internal
 
         private static CakeArguments CreateCakeArguments(IRemainingArguments remainingArguments, DefaultCommandSettings settings)
         {
-            var arguments = new Dictionary<string, List<string>>(StringComparer.OrdinalIgnoreCase);
-
-            // Keep the actual remaining arguments in the cake arguments
-            foreach (var group in remainingArguments.Parsed)
-            {
-                arguments[group.Key] = new List<string>();
-                foreach (var argument in group)
-                {
-                    arguments[group.Key].Add(argument);
-                }
-            }
-
-            // Fixes #3291, We have to add arguments manually which are defined within the DefaultCommandSettings type. Those are not considered "as remaining" because they could be parsed
-            const string targetArgumentName = "target";
-            if (!arguments.ContainsKey(targetArgumentName))
-            {
-                arguments[targetArgumentName] = new List<string>();
-            }
-            foreach (var target in settings.Targets)
-            {
-                arguments[targetArgumentName].Add(target);
-            }
-
-            var argumentLookUp = arguments.SelectMany(a => a.Value, Tuple.Create).ToLookup(a => a.Item1.Key, a => a.Item2);
-            return new CakeArguments(argumentLookUp);
+            return remainingArguments.ToCakeArguments(settings.Targets);
         }
 
         private void InstallTools(ServiceProvider provider)
